feat(redirect): disable follow redirect
Disable `followRedirects` by default and add a configuration options.
This can have an impact if a user was expecting a http -> https
redirection.
Feign request options was not inheriting from the okhttp configuration
for timeouts.

Fix BonitaClientIT as default layout and bootstrap theme are not
included in the bundle by default anymore.